moved loader types from ids to new loader_content module
